LPS-133888 Turn DDMDataProvider into a interface to avoid circular dependency
